refactor(otel-metrics): rename trace stats package
