+++ pkg/opentelemetry-mapping-go/otlp/metrics/tracestats/BUILD.bazel
@@ -2,13 +2,13 @@ load("@rules_go//go:def.bzl", "go_library")
 load("//bazel/rules/go:dd_agent_go_test.bzl", "dd_agent_go_test")
 
 go_library(
-    name = "sdktracestats",
+    name = "tracestats",
     srcs = ["sdk_trace_stats.go"],
-    importpath = "github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/otlp/metrics/sdktracestats",
+    importpath = "github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/otlp/metrics/tracestats",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/opentelemetry-mapping-go/otlp/attributes",
-        "//pkg/opentelemetry-mapping-go/otlp/metrics/sdktracestats/pb",
+        "//pkg/opentelemetry-mapping-go/otlp/metrics/tracestats/pb",
         "//pkg/trace/traceutil/normalize",
         "@io_opentelemetry_go_collector_pdata//pcommon",
         "@io_opentelemetry_go_collector_pdata//pmetric",
@@ -18,11 +18,11 @@ go_library(
 )
 
 dd_agent_go_test(
-    name = "sdktracestats_test",
+    name = "tracestats_test",
     srcs = ["sdk_trace_stats_test.go"],
-    embed = [":sdktracestats"],
+    embed = [":tracestats"],
     deps = [
-        "//pkg/opentelemetry-mapping-go/otlp/metrics/sdktracestats/pb",
+        "//pkg/opentelemetry-mapping-go/otlp/metrics/tracestats/pb",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
         "@io_opentelemetry_go_collector_pdata//pcommon",
